+++ app/src/main/java/dnd/jon/spellbook/CharacterSchoolDao.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface CharacterSchoolDao extends DAO<CharacterSchoolEntry> {
 
     @Query("SELECT * FROM character_schools WHERE (character_id = :characterID AND school_id = :schoolID)")
-    CharacterSchoolEntry getEntryByIds(int characterID, int schoolID);
+    CharacterSchoolEntry getEntryByIds(long characterID, long schoolID);
 
     @Query("SELECT * FROM schools INNER JOIN (SELECT school_id FROM character_schools WHERE character_id = :characterID) AS res ON schools.id = res.school_id")
-    List<School> getVisibleSchools(int characterID);
+    List<School> getVisibleSchools(long characterID);
 
 }
