+++ app/src/main/java/dnd/jon/spellbook/CharacterSourceDao.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface CharacterSourceDao extends DAO<CharacterSourceEntry> {
 
     @Query("SELECT * FROM character_sources WHERE (character_id = :characterID AND source_id = :sourceID)")
-    CharacterSourceEntry getEntryByIds(int characterID, int sourceID);
+    CharacterSourceEntry getEntryByIds(long characterID, long sourceID);
 
     @Query("SELECT * FROM sources INNER JOIN (SELECT source_id FROM character_sources WHERE character_id = :characterID) AS res ON sources.id = res.source_id")
-    List<Source> getVisibleSources(int characterID);
+    List<Source> getVisibleSources(long characterID);
 
 }
