+++ app/src/main/java/dnd/jon/spellbook/SourceDao.java
@@ -31,9 +31,9 @@ public interface SourceDao extends DAO<Source> {
     Source getSourceFromCode(String code);
 
     @Query("SELECT * FROM sources where id = :id")
-    Source getSourceByID(int id);
+    Source getSourceByID(long id);
 
     @Query("SELECT code FROM sources where id = :id")
-    String getSourceCodeByID(int id);
+    String getSourceCodeByID(long id);
 
 }
